// src/model/RelationModel.ts
/**
 * Created by guoyan on 2018/4/10.
 */
export class RelationModel {
  androidProjectId: string;
  androidProjectName: string;
  androidBranchId: string;
  androidBranchName: string;
  androidKey: string;

  iosProjectId: string;
  iosProjectName: string;
  iosBranchId: string;
  iosBranchName: string;
  iosKey: string;

}